refactor game page structure and add new widgets for betting interface
