+++ lib/features/card_game/presentation/game_page/screens/game_page.dart
@@ -1,8 +1,8 @@
-import 'package:aceplus/features/card_game/presentation/game_page/game_widgets/top_widget.dart';
+import 'package:aceplus/features/card_game/presentation/game_page/widgets/top_widget.dart';
 import 'package:flutter/material.dart';
-import '../game_widgets/bottom_widget.dart';
-import '../game_widgets/game_timer.dart';
-import '../game_widgets/table_widget.dart';
+import '../widgets/bottom_widget.dart';
+import '../widgets/timer_widget/game_timer.dart';
+import '../widgets/table_widget.dart';
 
 class GamePage extends StatelessWidget {
   const GamePage({super.key});
