+++ lib/shared/utils/constant.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/cupertino.dart';
 import 'package:google_fonts/google_fonts.dart';
